Feat : manual 기능 및 좌표 변환 기능 추가
manual 	  -> 전체화면 on, off을 위한 제스처 처리 구현
좌표 변환 -> picma 해상도와 pdf_viewer의 viewport() 크기 차이를 고려해
	     비율을 조정, 더 큰 비율로 좌표를 변환해 화면에 꽉 차도록 구현

+++ source/event_overlay_widget.cpp
@@ -33,6 +33,7 @@ Event_Overlay_Widget::~Event_Overlay_Widget(){
     }
     paths.clear();
 }
+
 void Event_Overlay_Widget::paintEvent(QPaintEvent *event){
     // painter 설정
     QPainter painter(this);
@@ -72,7 +73,7 @@ int Event_Overlay_Widget::get_paint_mode(){
     return current_paint_mode;
 }
 
-void Event_Overlay_Widget::set_pos(const int x, const int y){
+void Event_Overlay_Widget::set_pos(const qreal x, const qreal y){
     prev_pos = current_pos;
     current_pos.setX(x); current_pos.setY(y);
